Implement CutsetConditioningSolver

+++ src/main/java/org/jcsp/domains/Domain.java
@@ -11,7 +11,7 @@ import java.util.stream.Stream;
 public interface Domain {
     boolean contains(@Nullable Object value);
     boolean isEmpty();
-    long size();
+    int size();
     Stream<?> stream();
     Builder toBuilder();
